Split makeInputFile up, gave it its own file
