+++ code/drasil-code/drasil-code.cabal
@@ -18,7 +18,8 @@ library
     , Language.Drasil.Code.Imperative.Build.AST
     , Language.Drasil.Code.Imperative.Build.Import
     , Language.Drasil.Code.Imperative.Doxygen.Import
-    , Language.Drasil.Code.Imperative.Input
+    , Language.Drasil.Code.Imperative.ReadInput
+    , Language.Drasil.Code.Imperative.WriteInput
     , Language.Drasil.Code.Imperative.GOOL.Data
     , Language.Drasil.Code.Imperative.GOOL.Helpers
     , Language.Drasil.Code.Imperative.GOOL.LanguageRenderer
@@ -80,7 +81,8 @@ executable codegenTest
     , Language.Drasil.Code.Imperative.Build.AST
     , Language.Drasil.Code.Imperative.Build.Import
     , Language.Drasil.Code.Imperative.Doxygen.Import
-    , Language.Drasil.Code.Imperative.Input
+    , Language.Drasil.Code.Imperative.ReadInput
+    , Language.Drasil.Code.Imperative.WriteInput
     , Language.Drasil.Code.Imperative.GOOL.Data
     , Language.Drasil.Code.Imperative.GOOL.Helpers
     , Language.Drasil.Code.Imperative.GOOL.LanguageRenderer
